chore: moved Executor and Parsers (renamed to AllParsers) to ExecutionStage folder
